+++ mvc/src/main/java/nextstep/mvc/handler/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.handler;
 
 import nextstep.mvc.view.ModelAndView;
